refactor(i18n/data/scripts): use .lock instead of .ref + move related scripts to scripts/i18n
