+++ data/logement/logement.yaml
@@ -47,6 +47,7 @@ logement . notif minimum habitants:
     Le nombre d'habitants ne peut pas être nul. Vous êtes le premier habitant de votre logement.
 
 logement . surface:
+  titre: Surface
   question: Quelle est la surface de votre logement ?
   par défaut:
     variations:
@@ -78,6 +79,7 @@ logement . construction . annuelle et par surface:
   unité: kgCO2e/m2
 
 logement . appartement:
+  titre: Appartement
   question: Votre logement est-il un appartement ?
   par défaut: non
   description: |
