+++ data/transport/transport . avion.yaml
@@ -4,14 +4,14 @@ transport . avion:
   description: |
     ![](https://images.unsplash.com/photo-1525624286412-4099c83c1bc8?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=300&q=80)
 
-    L'empreinte de l'avion est conséquente. Il n'est pas forcément plus polluant *au kilomètre* qu'une voiture qui ne déplace qu'une personne, 
+    L'empreinte de l'avion est conséquente. Il n'est pas forcément plus polluant *au kilomètre* qu'une voiture qui ne déplace qu'une personne,
     mais la distance et surtout la distance parcourue en une heure (qu'on appelle couramment la vitesse), permet de changer de continent pour un week-end.
 
     Gardons aussi en tête que le Français moyen ne prend pas ou peu l'avion. L'essentiel de l'empreinte de l'aviation est donc concentrée sur 10% ou moins des Français.
 
   note: |
-    Nous avons fait le choix de demander le total d'heures de vol, plutôt que de demander à l'utilisateur de renseigner chacun de ses vols 
-    via l'aéroport de départ, d'arrivée et la classe économique. C'est une évolution possible par la suite, 
+    Nous avons fait le choix de demander le total d'heures de vol, plutôt que de demander à l'utilisateur de renseigner chacun de ses vols
+    via l'aéroport de départ, d'arrivée et la classe économique. C'est une évolution possible par la suite,
     voir [ce simulateur](https://futur.eco/simulateur/transport/avion/impact) où une partie du travail est déjà faite.
 
     A noter aussi que les escales entrainent des décollages supplémentaires, qui consomment plus de carburant qu'une vitesse de croisière.
@@ -19,7 +19,7 @@ transport . avion:
 
     Pour un modèle plus détaillé de l'empreinte de l'avion, nous avons trouvé deux sources.
 
-    L'algorithme complet de l'appli North, disponible [ici](https://github.com/tmrowco/northapp-contrib/blob/master/co2eq/flights/index.js).  
+    L'algorithme complet de l'appli North, disponible [ici](https://github.com/tmrowco/northapp-contrib/blob/master/co2eq/flights/index.js).
 
     Le pseudo-algorithme de myclimate.org, ci-dessous.
 
@@ -33,7 +33,7 @@ transport . avion:
 transport . avion . usager:
   question: Prenez-vous l'avion ?
   description: |
-    Si vous ne prenez pas l'avion tous les ans mais que vous comptez continuer à le prendre, répondez "oui" et estimez grossièrement le nombre d'heures de vol ramené à l'année. 
+    Si vous ne prenez pas l'avion tous les ans mais que vous comptez continuer à le prendre, répondez "oui" et estimez grossièrement le nombre d'heures de vol ramené à l'année.
 
     Ce n'est pas très rigoureux, mais l'avion a une telle empreinte sur le climat qu'il est important de prendre cela en compte.
 
@@ -123,7 +123,7 @@ transport . avion . long courrier . empreinte par km:
 transport . avion . long courrier . heures de vol:
   question: Combien d'heures par an voyagez-vous au-delà en vol long courrier ?
   description: |
-    Comptez les heures que vous avez passé dans un avion pour des vols d'une distance supérieure à 3500km, c'est-à-dire au-delà de l'Europe et du pourtour méditerranéen. 
+    Comptez les heures que vous avez passé dans un avion pour des vols d'une distance supérieure à 3500km, c'est-à-dire au-delà de l'Europe et du pourtour méditerranéen.
 
     Si vous faites des escales, additionnez simplement les durées de chaque vol.
