+++ package.json
@@ -11,11 +11,11 @@
 		"compile:personas": "node scripts/personasToJSON.js",
 		"compile:rules": "node scripts/rulesToJSON.js",
 		"test": "echo \"Error: no test specified\" && exit 1",
-		"translate": "node scripts/translate-rules.js && node scripts/translate-personas.js",
-		"translate:personas:check": "node scripts/check-personas.js ",
-		"translate:personas": "node scripts/translate-personas.js",
-		"translate:rules:check": "node scripts/check-translation.js",
-		"translate:rules": "node scripts/translate-rules.js"
+		"translate": "node scripts/i18n/translate-rules.js && node scripts/i18n/translate-personas.js",
+		"translate:personas:check": "node scripts/i18n/check-personas.js ",
+		"translate:personas": "node scripts/i18n/translate-personas.js",
+		"translate:rules:check": "node scripts/i18n/check-translation.js",
+		"translate:rules": "node scripts/i18n/translate-rules.js"
 	},
 	"repository": {
 		"type": "git",
