+++ scripts/i18n/addTranslationToBasePersonas.js
